espd/requirements.txt
absl-py==1.3.0
addict==2.4.0
alabaster==0.7.8
albumentations==1.3.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==19.3.0
Automat==0.8.0
Babel==2.10.3
backcall==0.2.0
bcrypt==4.0.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
cachetools==5.2.0
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
click==8.0.4
cloud-init==22.2
cloudpickle==2.2.0
colorama==0.4.3
coloredlogs==15.0.1
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.5
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
deprecation==2.1.0
distro==1.7.0
distro-info===0.23ubuntu1
dm-tree==0.1.8
docker==6.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
entrypoints==0.3
executing==1.1.0
fastjsonschema==2.16.2
feedparser==5.2.1
fire==0.5.0
flatbuffers==22.10.26
fonttools==4.37.3
gast==0.4.0
gitsome==0.8.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
httplib2==0.14.0
humanfriendly==10.0
hyperlink==19.0.0
idna==2.8
imageio==2.22.4
imagesize==1.2.0
importlib-metadata==4.12.0
incremental==16.10.1
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyter_client==7.3.5
jupyter_packaging==0.12.3
jupyterlab==3.4.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.15.1
keras==2.11.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.4
labelme2coco==0.2.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==14.0.6
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
more-itertools==4.2.0
mpmath==1.2.1
nbclassic==0.4.3
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.6.0
nest-asyncio==1.5.5
netifaces==0.10.4
networkx==2.8.8
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.22.4
numpydoc==0.7.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.1.0
olefile==0.46
onnx==1.12.0
onnxruntime==1.13.1
open3d==0.13.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==20.3
pandas==1.5.0
pandocfilters==1.5.0
paramiko==2.11.0
parso==0.8.3
pbr==5.10.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.2.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pybboxes==0.1.6
pybind11==2.10.1
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pycurl==7.43.0.2
Pygments==2.13.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
pytorch-metric-learning==1.6.3
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qudida==0.0.4
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.9
sahi==0.11.10
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
SecretStorage==2.3.1
Send2Trash==1.8.0
service-identity==18.1.0
setproctitle==1.1.10
shapely==2.0.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
sos==4.3
sounddevice==0.4.5
soupsieve==2.3.2.post1
Sphinx==1.8.5
ssh-import-id==5.10
stack-data==0.5.1
sympy==1.11.1
systemd-python==234
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons==0.18.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-probability==0.19.0
termcolor==2.1.0
terminado==0.15.0
terminaltables==3.1.10
testresources==2.0.1
texttable==1.6.4
tflite-support==0.4.3
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.11.4
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
Twisted==18.9.0
typeguard==2.13.3
typing_extensions==4.4.0
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.12
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.59.0
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
xkit==0.0.0
xonsh==0.9.13
yapf==0.32.0
zipp==1.0.0
zope.interface==4.7.1
